Move Rita exit dashboard into it's own file
