+++ rita_bin/src/exit.rs
@@ -39,10 +39,10 @@ use rita_common::rita_loop::write_to_disk::SettingsOnDisk;
 use rita_common::usage_tracker::save_usage_on_shutdown;
 use rita_common::utils::apply_babeld_settings_defaults;
 use rita_common::utils::env_vars_contains;
+use rita_exit::dashboard::start_rita_exit_dashboard;
 use rita_exit::operator_update::update_loop::start_operator_update_loop;
 use rita_exit::rita_loop::start_rita_exit_endpoints;
 use rita_exit::rita_loop::start_rita_exit_loop;
-use rita_exit::start_rita_exit_dashboard;
 use rita_exit::{get_exit_usage, Args};
 use settings::exit::RitaExitSettingsStruct;
 use settings::save_settings_on_shutdown;
